requirements.txt
# Docildos - Requirements Universal
# Este arquivo contém todas as dependências do projeto (Backend + AI Service)

# ============================================================================
# Web Framework
# ============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ============================================================================
# Database
# ============================================================================
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# ============================================================================
# AI / LLM
# ============================================================================
langgraph==0.0.20
langchain==0.1.6
langchain-openai==0.0.5
langchain-ollama==0.1.0
openai==1.12.0

# ============================================================================
# Data Validation & Settings
# ============================================================================
pydantic==2.5.3
pydantic-settings==2.1.0

# ============================================================================
# HTTP Client & Utilities
# ============================================================================
httpx==0.26.0
python-dotenv==1.0.0

